refactor so docs can import
